backend/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/dev.in
#
asgiref==3.8.1
    # via
    #   -r requirements/test.txt
    #   django
    #   django-cors-headers
astroid==3.2.4
    # via
    #   -r requirements/test.txt
    #   pylint
black==24.8.0
    # via -r requirements/dev.in
build==1.2.2
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   pip-tools
coverage[toml]==7.6.1
    # via
    #   -r requirements/test.txt
    #   pytest-cov
dill==0.3.8
    # via
    #   -r requirements/test.txt
    #   pylint
django==4.2
    # via
    #   -r requirements/test.txt
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.4.0
    # via -r requirements/test.txt
django-environ==0.10.0
    # via -r requirements/test.txt
djangorestframework==3.14.0
    # via
    #   -r requirements/test.txt
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements/test.txt
exceptiongroup==1.2.2
    # via
    #   -r requirements/test.txt
    #   pytest
importlib-metadata==8.4.0
    # via build
iniconfig==2.0.0
    # via
    #   -r requirements/test.txt
    #   pytest
isort==5.13.2
    # via
    #   -r requirements/test.txt
    #   pylint
mccabe==0.7.0
    # via
    #   -r requirements/test.txt
    #   pylint
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via
    #   -r requirements/test.txt
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements/dev.in
platformdirs==4.3.2
    # via
    #   -r requirements/test.txt
    #   black
    #   pylint
pluggy==1.5.0
    # via
    #   -r requirements/test.txt
    #   pytest
pyjwt==2.9.0
    # via
    #   -r requirements/test.txt
    #   djangorestframework-simplejwt
pylint==3.2.7
    # via
    #   -r requirements/test.txt
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.5
    # via -r requirements/test.txt
pylint-plugin-utils==0.8.2
    # via
    #   -r requirements/test.txt
    #   pylint-django
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==8.3.3
    # via
    #   -r requirements/test.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.txt
pytest-django==4.9.0
    # via -r requirements/test.txt
pytz==2024.1
    # via
    #   -r requirements/test.txt
    #   djangorestframework
sqlparse==0.5.1
    # via
    #   -r requirements/test.txt
    #   django
tomli==2.0.1
    # via
    #   -r requirements/test.txt
    #   black
    #   build
    #   coverage
    #   pip-tools
    #   pylint
    #   pytest
tomlkit==0.13.2
    # via
    #   -r requirements/test.txt
    #   pylint
typing-extensions==4.12.2
    # via
    #   -r requirements/test.txt
    #   asgiref
    #   astroid
    #   black
    #   pylint
wheel==0.44.0
    # via pip-tools
zipp==3.20.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
